added connection to postgresql for faster queries
